requirements.txt
torch
pyyaml
spacy
numpy
scipy
sklearn
tensorboard